+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.1
 * Fix flush agent and replication agent initialisation parameters mix up
+* Add package upload_wait_until_ready and install_wait_until_ready methods
 
 ### 1.0.0
 * Fix replication agent creation failure #1